% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{The number of a state in the US as used in the FARS dataset
Should be numeric or integer.}

\item{year}{The year of analysis (numeric or integer)}
}
\value{
a graphical overview of fatalities on a map in a particular time period. 
Returns an error if the state or year do not exist in the data set.
}
\description{
This function takes a state number and set of years as input and shows 
an overview of fatalities on a map in that particular time period.

Uses function make_filename and fars_read from the current package. 
Removes coordinate outliers - longitude values greater than 900 
and latitude values greater than 90 are removed.
}
\examples{
\dontrun{
fars_map_state(45, 2015)

# Results in an error
fars_map_state(45, 2016)
fars_map_state(60, 2015)
}

}
