% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{fars_read}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character object which corresponds to a valid path of the data file.
In case the file does not exist an error message is produced and execution stops.}
}
\value{
The function returns a dataframe based on the CSV file.
}
\description{
Reads in file to data variable using the read_csv function
and creates a dataframe summarizing the contents.
}
\details{
Building R Packages - Week 2 Assignment
filename: fars_functions.R

These functions read in data taken from the US National Highway Traffic Safety Administration's
\href{https://www.nhtsa.gov/Data/Fatality-Analysis-Reporting-System-(FARS)}{Fatality Analysis Reporting System}
}
\examples{
\dontrun{
accident_2015 <- fars_read(".inst/extdata/accident_2015.csv.bz2")
}
}
