% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector or list of years in numeric or integer format}
}
\value{
Returns a list of dataframe with the same number of rows
as the data in "accident_<year>.csv.bz2" files sorted by year and MONTH.

If any of the objects requested via input is not available as a year file
or is not coercible to integer an "invalid year" error message returns.
}
\description{
The function accepts a vector or list of years and returns a list of dataframe
with MONTH and year columns based on data in "accident_<year>.csv.bz2
files. The files need to be located in the working directory.
}
\examples{
\dontrun{
fars_read_years(2013:2015)
fars_read_years(list(2013, 2014))

# Results in a warning
fars_read_years(2016)
}
}
