% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector or list of years (numeric or integer) to 
read in and summarize

The function will take in a vector of years and read in the data using 
the fars_summarize_years function, it then binds these rows together and 
groups by the year and MONTH column creating a count column: n. 
The data is then converted from a long format to a wide format using 
the spread function in tidyr.}
}
\value{
a data.frame of summarized data which is converted to a wide format
}
\description{
takes a list of years and reads them in using fars_read_years 
it then binds these dataframe together and summarizes the data.
}
\examples{
\donttest{
 library(farspkg)
 fars_summarize_years("2015")
 fars_summarize_years(c(2013.0,2014))
 }
}
